Attempt to listen to server's camera broadcast
It just inserts an empty video element, for some reason.

Also this includes hardcoded server and client URLs that will need to be
made dynamic, but I ripped out a lot of ENV stuff already so I'll do
that in a bit (plus don't know how the actual deployment will look yet).